+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `draw_vertexlabels`, `draw_edgelabels`, `draw_facelabels`, `draw_vertexnormals`, and `draw_facenormals` to `compas_blender.artists.MeshArtist`.
-* Added `Plane.offset`
+* Added optional `triangulated` flag to `to_vertices_and_faces` of all shapes.
+* Added `compas.geometry.Geometry` base class.
+* Added `__add__`, `__sub__`, `__and__` to `compas.geometry.Shape` for boolean operations using binary operators.
+* Added `is_closed` to `compas.geometry.Polyhedron`.
+* Added `Plane.offset`.
 * Added `compas.artists.Artist`.
 * Added pluggable `compas.artists.new_artist`.
 * Added plugin `compas_rhino.artists.new_artist_rhino`.
@@ -24,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed bug in `compas_blender.draw_texts`.
 * Changed `compas_rhino.artists.BaseArtist` to `compas_rhino.artists.RhinoArtist`.
 * Changed `compas_blender.artists.BaseArtist` to `compas_blender.artists.BlenderArtist`.
+* Changed default resolution for shape discretisation to 16 for both u and v where relevant.
+* Changed base class of `compas.geometry.Primitive` and `compas.geometry.Shape` to `compas.geometry.Geometry`.
 * `compas_blender.artists.RobotModelArtist.collection` can be assigned as a Blender collection or a name.
 * Generalized the parameter `color` of `compas_blender.draw_texts` and various label drawing methods.